+++ NEWS.md
@@ -5,6 +5,7 @@
 * Support `projpred` version 2.0 for variable selection in generalized
 linear and additive multilevel models thanks to Alejandro Catalina.
 * Support `by` variables in multi-membership terms.
+* Use Bayesian bootstrap in `loo_R2`.
 
 ### Bug Fixes